Switch to chicago_formatter to override, update to match hyrax 3, adjust doi inclusion to add in space when needed without having to override more methods like before
